refactor : change entity and vo folder called domain into model
- 도메인 페키지와 이름이 겹치는 것의 오는 혼란을 없엡니다
