+++ greendar/src/main/java/greendar/domain/privatetodo/api/PrivateTodoApi.java
@@ -1,10 +1,10 @@
 package greendar.domain.privatetodo.api;
 
 import greendar.domain.member.application.MemberService;
-import greendar.domain.member.domain.Member;
+import greendar.domain.member.model.Member;
 import greendar.domain.privatetodo.application.PrivateTodoService;
 import greendar.domain.privatetodo.dto.PrivateTodoDtos.DailyAchievementRatio;
-import greendar.domain.privatetodo.domain.PrivateTodo;
+import greendar.domain.privatetodo.model.PrivateTodo;
 import greendar.domain.privatetodo.dto.PrivateTodoDtos.PrivateTodoCompletePutRequestDto;
 import greendar.domain.privatetodo.dto.PrivateTodoDtos.PrivateTodoPostRequestDto;
 import greendar.domain.privatetodo.dto.PrivateTodoDtos.PrivateTodoResponse;
